refactor write.go
