+++ executor/load_data.go
@@ -1,0 +1,371 @@
+// Copyright 2018 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package executor
+
+import (
+	"bytes"
+	"fmt"
+	"strings"
+
+	"github.com/juju/errors"
+	"github.com/pingcap/tidb/ast"
+	"github.com/pingcap/tidb/sessionctx"
+	"github.com/pingcap/tidb/table"
+	"github.com/pingcap/tidb/types"
+	"github.com/pingcap/tidb/util/chunk"
+	log "github.com/sirupsen/logrus"
+	"golang.org/x/net/context"
+)
+
+// LoadDataInfo saves the information of loading data operation.
+type LoadDataInfo struct {
+	row       []types.Datum
+	insertVal *InsertValues
+
+	Path       string
+	Table      table.Table
+	FieldsInfo *ast.FieldsClause
+	LinesInfo  *ast.LinesClause
+	Ctx        sessionctx.Context
+	columns    []*table.Column
+}
+
+// SetMaxRowsInBatch sets the max number of rows to insert in a batch.
+func (e *LoadDataInfo) SetMaxRowsInBatch(limit uint64) {
+	e.insertVal.maxRowsInBatch = limit
+}
+
+// getValidData returns prevData and curData that starts from starting symbol.
+// If the data doesn't have starting symbol, prevData is nil and curData is curData[len(curData)-startingLen+1:].
+// If curData size less than startingLen, curData is returned directly.
+func (e *LoadDataInfo) getValidData(prevData, curData []byte) ([]byte, []byte) {
+	startingLen := len(e.LinesInfo.Starting)
+	if startingLen == 0 {
+		return prevData, curData
+	}
+
+	prevLen := len(prevData)
+	if prevLen > 0 {
+		// starting symbol in the prevData
+		idx := strings.Index(string(prevData), e.LinesInfo.Starting)
+		if idx != -1 {
+			return prevData[idx:], curData
+		}
+
+		// starting symbol in the middle of prevData and curData
+		restStart := curData
+		if len(curData) >= startingLen {
+			restStart = curData[:startingLen-1]
+		}
+		prevData = append(prevData, restStart...)
+		idx = strings.Index(string(prevData), e.LinesInfo.Starting)
+		if idx != -1 {
+			return prevData[idx:prevLen], curData
+		}
+	}
+
+	// starting symbol in the curData
+	idx := strings.Index(string(curData), e.LinesInfo.Starting)
+	if idx != -1 {
+		return nil, curData[idx:]
+	}
+
+	// no starting symbol
+	if len(curData) >= startingLen {
+		curData = curData[len(curData)-startingLen+1:]
+	}
+	return nil, curData
+}
+
+// getLine returns a line, curData, the next data start index and a bool value.
+// If it has starting symbol the bool is true, otherwise is false.
+func (e *LoadDataInfo) getLine(prevData, curData []byte) ([]byte, []byte, bool) {
+	startingLen := len(e.LinesInfo.Starting)
+	prevData, curData = e.getValidData(prevData, curData)
+	if prevData == nil && len(curData) < startingLen {
+		return nil, curData, false
+	}
+
+	prevLen := len(prevData)
+	terminatedLen := len(e.LinesInfo.Terminated)
+	curStartIdx := 0
+	if prevLen < startingLen {
+		curStartIdx = startingLen - prevLen
+	}
+	endIdx := -1
+	if len(curData) >= curStartIdx {
+		endIdx = strings.Index(string(curData[curStartIdx:]), e.LinesInfo.Terminated)
+	}
+	if endIdx == -1 {
+		// no terminated symbol
+		if len(prevData) == 0 {
+			return nil, curData, true
+		}
+
+		// terminated symbol in the middle of prevData and curData
+		curData = append(prevData, curData...)
+		endIdx = strings.Index(string(curData[startingLen:]), e.LinesInfo.Terminated)
+		if endIdx != -1 {
+			nextDataIdx := startingLen + endIdx + terminatedLen
+			return curData[startingLen : startingLen+endIdx], curData[nextDataIdx:], true
+		}
+		// no terminated symbol
+		return nil, curData, true
+	}
+
+	// terminated symbol in the curData
+	nextDataIdx := curStartIdx + endIdx + terminatedLen
+	if len(prevData) == 0 {
+		return curData[curStartIdx : curStartIdx+endIdx], curData[nextDataIdx:], true
+	}
+
+	// terminated symbol in the curData
+	prevData = append(prevData, curData[:nextDataIdx]...)
+	endIdx = strings.Index(string(prevData[startingLen:]), e.LinesInfo.Terminated)
+	if endIdx >= prevLen {
+		return prevData[startingLen : startingLen+endIdx], curData[nextDataIdx:], true
+	}
+
+	// terminated symbol in the middle of prevData and curData
+	lineLen := startingLen + endIdx + terminatedLen
+	return prevData[startingLen : startingLen+endIdx], curData[lineLen-prevLen:], true
+}
+
+// InsertData inserts data into specified table according to the specified format.
+// If it has the rest of data isn't completed the processing, then is returns without completed data.
+// If the number of inserted rows reaches the batchRows, then the second return value is true.
+// If prevData isn't nil and curData is nil, there are no other data to deal with and the isEOF is true.
+func (e *LoadDataInfo) InsertData(prevData, curData []byte) ([]byte, bool, error) {
+	// TODO: support enclosed and escape.
+	if len(prevData) == 0 && len(curData) == 0 {
+		return nil, false, nil
+	}
+
+	var line []byte
+	var isEOF, hasStarting, reachLimit bool
+	if len(prevData) > 0 && len(curData) == 0 {
+		isEOF = true
+		prevData, curData = curData, prevData
+	}
+	rows := make([][]types.Datum, 0, e.insertVal.maxRowsInBatch)
+	for len(curData) > 0 {
+		line, curData, hasStarting = e.getLine(prevData, curData)
+		prevData = nil
+
+		// If it doesn't find the terminated symbol and this data isn't the last data,
+		// the data can't be inserted.
+		if line == nil && !isEOF {
+			break
+		}
+		// If doesn't find starting symbol, this data can't be inserted.
+		if !hasStarting {
+			if isEOF {
+				curData = nil
+			}
+			break
+		}
+		if line == nil && isEOF {
+			line = curData[len(e.LinesInfo.Starting):]
+			curData = nil
+		}
+
+		cols, err := GetFieldsFromLine(line, e.FieldsInfo)
+		if err != nil {
+			return nil, false, errors.Trace(err)
+		}
+		rows = append(rows, e.colsToRow(cols))
+		e.insertVal.rowCount++
+		if e.insertVal.maxRowsInBatch != 0 && e.insertVal.rowCount%e.insertVal.maxRowsInBatch == 0 {
+			reachLimit = true
+			log.Infof("This insert rows has reached the batch %d, current total rows %d",
+				e.insertVal.maxRowsInBatch, e.insertVal.rowCount)
+			break
+		}
+	}
+	rows, err := batchMarkDupRows(e.Ctx, e.Table, rows)
+	if err != nil {
+		return nil, reachLimit, errors.Trace(err)
+	}
+	for _, row := range rows {
+		e.insertData(row)
+	}
+	if e.insertVal.lastInsertID != 0 {
+		e.insertVal.ctx.GetSessionVars().SetLastInsertID(e.insertVal.lastInsertID)
+	}
+
+	return curData, reachLimit, nil
+}
+
+func (e *LoadDataInfo) colsToRow(cols []string) types.DatumRow {
+	for i := 0; i < len(e.row); i++ {
+		if i >= len(cols) {
+			e.row[i].SetString("")
+			continue
+		}
+		e.row[i].SetString(cols[i])
+	}
+	row, err := e.insertVal.fillRowData(e.columns, e.row)
+	if err != nil {
+		warnLog := fmt.Sprintf("Load Data: insert data:%v failed:%v", e.row, errors.ErrorStack(err))
+		e.insertVal.handleLoadDataWarnings(err, warnLog)
+		return nil
+	}
+	return row
+}
+
+func (e *LoadDataInfo) insertData(row types.DatumRow) {
+	if row == nil {
+		return
+	}
+	_, err := e.Table.AddRecord(e.insertVal.ctx, row, false)
+	if err != nil {
+		warnLog := fmt.Sprintf("Load Data: insert data:%v failed:%v", row, errors.ErrorStack(err))
+		e.insertVal.handleLoadDataWarnings(err, warnLog)
+	}
+}
+
+// LoadData represents a load data executor.
+type LoadData struct {
+	baseExecutor
+
+	IsLocal      bool
+	loadDataInfo *LoadDataInfo
+}
+
+// Next implements the Executor Next interface.
+func (e *LoadData) Next(ctx context.Context, chk *chunk.Chunk) error {
+	chk.Reset()
+	// TODO: support load data without local field.
+	if !e.IsLocal {
+		return errors.New("Load Data: don't support load data without local field")
+	}
+	// TODO: support lines terminated is "".
+	if len(e.loadDataInfo.LinesInfo.Terminated) == 0 {
+		return errors.New("Load Data: don't support load data terminated is nil")
+	}
+
+	sctx := e.loadDataInfo.insertVal.ctx
+	val := sctx.Value(LoadDataVarKey)
+	if val != nil {
+		sctx.SetValue(LoadDataVarKey, nil)
+		return errors.New("Load Data: previous load data option isn't closed normal")
+	}
+	if e.loadDataInfo.Path == "" {
+		return errors.New("Load Data: infile path is empty")
+	}
+	sctx.SetValue(LoadDataVarKey, e.loadDataInfo)
+
+	return nil
+}
+
+// Close implements the Executor Close interface.
+func (e *LoadData) Close() error {
+	return nil
+}
+
+// Open implements the Executor Open interface.
+func (e *LoadData) Open(ctx context.Context) error {
+	return nil
+}
+
+// NewLoadDataInfo returns a LoadDataInfo structure, and it's only used for tests now.
+func NewLoadDataInfo(ctx sessionctx.Context, row []types.Datum, tbl table.Table, cols []*table.Column) *LoadDataInfo {
+	insertVal := &InsertValues{baseExecutor: newBaseExecutor(ctx, nil, "InsertValues"), Table: tbl}
+	return &LoadDataInfo{
+		row:       row,
+		insertVal: insertVal,
+		Table:     tbl,
+		Ctx:       ctx,
+		columns:   cols,
+	}
+}
+
+// GetFieldsFromLine splits line according to fieldsInfo, this function is exported for testing.
+func GetFieldsFromLine(line []byte, fieldsInfo *ast.FieldsClause) ([]string, error) {
+	var sep []byte
+	if fieldsInfo.Enclosed != 0 {
+		if line[0] != fieldsInfo.Enclosed || line[len(line)-1] != fieldsInfo.Enclosed {
+			return nil, errors.Errorf("line %s should begin and end with %c", string(line), fieldsInfo.Enclosed)
+		}
+		line = line[1 : len(line)-1]
+		sep = make([]byte, 0, len(fieldsInfo.Terminated)+2)
+		sep = append(sep, fieldsInfo.Enclosed)
+		sep = append(sep, fieldsInfo.Terminated...)
+		sep = append(sep, fieldsInfo.Enclosed)
+	} else {
+		sep = []byte(fieldsInfo.Terminated)
+	}
+	rawCols := bytes.Split(line, sep)
+	cols := escapeCols(rawCols)
+	return cols, nil
+}
+
+func escapeCols(strs [][]byte) []string {
+	ret := make([]string, len(strs))
+	for i, v := range strs {
+		output := escape(v)
+		ret[i] = string(output)
+	}
+	return ret
+}
+
+// escape handles escape characters when running load data statement.
+// TODO: escape need to be improved, it should support ESCAPED BY to specify
+// the escape character and handle \N escape.
+// See http://dev.mysql.com/doc/refman/5.7/en/load-data.html
+func escape(str []byte) []byte {
+	pos := 0
+	for i := 0; i < len(str); i++ {
+		c := str[i]
+		if c == '\\' && i+1 < len(str) {
+			c = escapeChar(str[i+1])
+			i++
+		}
+
+		str[pos] = c
+		pos++
+	}
+	return str[:pos]
+}
+
+func escapeChar(c byte) byte {
+	switch c {
+	case '0':
+		return 0
+	case 'b':
+		return '\b'
+	case 'n':
+		return '\n'
+	case 'r':
+		return '\r'
+	case 't':
+		return '\t'
+	case 'Z':
+		return 26
+	case '\\':
+		return '\\'
+	}
+	return c
+}
+
+// loadDataVarKeyType is a dummy type to avoid naming collision in context.
+type loadDataVarKeyType int
+
+// String defines a Stringer function for debugging and pretty printing.
+func (k loadDataVarKeyType) String() string {
+	return "load_data_var"
+}
+
+// LoadDataVarKey is a variable key for load data.
+const LoadDataVarKey loadDataVarKeyType = 0
